Moved Country view model factory in to a separate class file
