+++ app/src/main/java/com/codetutor/countryinfoapp/viewmodel/CountryViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.codetutor.countryinfoapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.codetutor.countryinfoapp.repository.CountryRepository
+
+class CountryViewModelFactory(private val repository: CountryRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(CountryViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return CountryViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
